Minor code refacoring

+++ Telemetry/code/MicroSD_SDIO/Core/Inc/sensors/ABS.h
@@ -23,7 +23,7 @@ typedef struct ABS{
 	char path[20]; // path of file to write;
 	int dataReady; // flag to check if data is ready to read and write to file
 	int saveRate;
-	int data;
+	uint16_t data;
 	int ID;
 	//ABS data asynchronous
 	int timeToZeroSpeed; //after that time if not new input we assume car is not moving
